fix: move extractEarliestYear to client-safe module to fix webpack fs error
risks-table.tsx ("use client") imported risks-sort.ts → risk-utils.ts →
directory-utils.ts → kb.ts which uses fs, causing "Module not found: Can't
resolve 'fs'" on Next.js build. Moved extractEarliestYear to risk-constants.ts
which has no server dependencies.
